user-service/requirements.txt
flask
flask_sqlalchemy
flask_redis
psycopg2-binary
python-dotenv
